feat(builtin): add echo for text output
Implemented the `echo` command, allowing users to output text to
the terminal. Currently, the command supports basic text output and
options, but variable expansion is not yet implemented.

Refs: #10

+++ builtin/cd.c
@@ -100,7 +100,7 @@ int builtin_cd(Command *cmd) {
   int print_new_dir = 0;
   const char *dir = NULL;
 
-  for (int i = 1; i < cmd->argv.size; i++) {
+  for (size_t i = 1; i < cmd->argv.size; i++) {
     if (strcmp(cmd->argv.data[i], "-L") == 0) {
       follow_symlinks = 1;
       physical = 0;
